moved mediator up
